update cmocka fetching

+++ src/platform/vt_dsc_custom.c
@@ -28,7 +28,7 @@ uint32_t _vt_dsc_gpio_turn_off(GPIO_PORT_TYPEDEF* GPIOx, GPION_PIN_TYPEDEF GPIO_
     return VT_CUSTOM_IMPLEMENTATION_NOT_PROVIDED;
 }
 
-uint32_t _vt_dsc_adc_read(ADC_CONTROLLER_TYPEDEF* ADC_Controller, ADC_CHANNEL_TYPEDEF ADC_Channel, float* value)
+uint32_t _vt_dsc_adc_read(ADC_CONTROLLER_TYPEDEF* ADC_Controller, ADC_CHANNEL_TYPEDEF ADC_Channel, uint32_t* value)
 {
     return VT_CUSTOM_IMPLEMENTATION_NOT_PROVIDED;
 }
